♻️  refactor(fe): 절대 경로 별칭 추가 및 폴더 구조 재정리
