+++ src/pages/Lecture/components/LectureIntroSection.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Intro from '../styles/LectureIntroSection.styled';
 import { StyledBaseSection as S } from '../styles/LectureDetailPage.styled';
-import { useLectureContext } from '../../../context/LectureContext';
+import { useLectureContext } from '@/pages/Lecture/contexts/LectureContext';
 
 const LectureIntroSection = React.forwardRef<HTMLElement>((_, ref) => {
   const { data } = useLectureContext();
